Related to #3205 

Tested this against https://github.com/harsh8398/fastapi-dependency-issue