@Kludex 2022-07-11
If a new `CapacityLimiter` is created every time, then the token can be 1... Can't it?

The inf would make sense for a global limiter... Isn't it the case?