@adriangb 2022-07-11
Yeah I guess it can be 1. Or we can make it a global. Either way.