@tiangolo 2022-09-04
Niceee! :nerd_face: :rocket:

I tested this with the example code in: https://github.com/tiangolo/fastapi/issues/3205

I used `wrk` and I ran twice:

```console
$ wrk -c 100 -t 100 --timeout 60 "http://127.0.0.1:8000/"
```

## Without the fix

Without the fix, I got first:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.40s     2.04s    8.05s    63.64%
    Req/Sec     0.00      0.00     0.00    100.00%
  22 requests in 10.09s, 3.05KB read
Requests/sec:      2.18
Transfer/sec:     309.54B
```

And then the second run I got:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.00us    0.00us   0.00us    -nan%
    Req/Sec     0.00      0.00     0.00      -nan%
  0 requests in 10.08s, 0.00B read
Requests/sec:      0.00
Transfer/sec:       0.00B
```

...this seems to reflect the deadlock. :tada: (I'm celebrating being able to see something broken, it's sometimes hard to see the broken part :sweat_smile: )

After that, I `Ctrl+C` the Uvicorn process, and it hung in there, at:

```
^C
INFO:     Shutting down
INFO:     Waiting for background tasks to complete. (CTRL+C to force quit)
```

After 7 minutes I force-killed the process. So, indeed, deadlock.

## With the fix, outside

With the fix, putting a single global `CapacityLimiter` outside of the function.

First I got:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.97s     3.07s   10.02s    67.50%
    Req/Sec     0.00      0.00     0.00    100.00%
  40 requests in 10.09s, 5.55KB read
Requests/sec:      3.96
Transfer/sec:     562.78B
```

For the second run I got:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.31s     2.45s    9.64s    73.08%
    Req/Sec     0.00      0.00     0.00    100.00%
  26 requests in 10.09s, 3.61KB read
Requests/sec:      2.58
Transfer/sec:     365.96B
```

Third run:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     6.90s     2.76s    9.58s    83.33%
    Req/Sec     0.00      0.00     0.00    100.00%
  12 requests in 10.08s, 1.66KB read
Requests/sec:      1.19
Transfer/sec:     169.06B
```

Not sure why it keeps reducing the number of requests. It seems something would be still wrong, but not sure what. :pensive: 

After that, and after I wrote this, I `Ctrl+C` the Uvicorn process, it finished immediately. So it seems it handled the deadlock.

I ran the whole thing again, and I `Ctrl+C` the Uvicorn process immediately after the 3 runs, and then it waited for a bit at:

```
^C
INFO:     Shutting down
INFO:     Waiting for background tasks to complete. (CTRL+C to force quit)
```

But after 50 seconds it terminated normally.

## With the fix, local

Keeping the local `CapacityLimiter` created every time, as in this PR.

The first run I got:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.71s     2.90s   10.06s    65.00%
    Req/Sec     0.00      0.00     0.00    100.00%
  40 requests in 10.08s, 5.55KB read
Requests/sec:      3.97
Transfer/sec:     563.34B
```

The second run I got:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.26s     2.50s    8.48s    80.00%
    Req/Sec     0.00      0.00     0.00    100.00%
  5 requests in 10.09s, 710.00B read
Requests/sec:      0.50
Transfer/sec:      70.37B
```

And the third run I got:

```
Running 10s test @ http://127.0.0.1:8000/
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     5.52s     2.99s    9.27s    50.00%
    Req/Sec     0.00      0.00     0.00    100.00%
  4 requests in 10.10s, 568.00B read
Requests/sec:      0.40
Transfer/sec:      56.25B
```

The results are similar to having a global `CapacityLimiter` than having a local one. The 5 or 4 requests are comparable to some of the other tests I ran with the global `CapacityLimiter`. It's in the same ballpark.

Terminating the process also worked as expected instead of hanging there in a deadlock.

For some reason, it's still being able to handle fewer requests every time. Maybe it's just that it's trying to read the file every time or something like that and there's a bottleneck in that, just in the file system. I'm not sure about that.

---

Checking the AnyIO docs I see that:

> Capacity limiters are like semaphores except that a single borrower (the current task by default) can only hold a single token at a time.

Ref: https://anyio.readthedocs.io/en/stable/synchronization.html

Then I'm not sure if it could be problematic if the same thread worker tried to get the limiter again, and if that could block with the other thing running on the same thread that already got the limiter token.

Until we figure this out, I'll leave the local `CapacityLimiter` (as in this PR) to be safe (although in my local tests it didn't seem to affect, but not sure why).

But anyway, it seems that this clearly improves the situation.

I wish there was a way to test this, in pytest. But not sure how to replicate the whole problem and situation in a feasible way (not involving a whole setup for "benchmarks" with `ab`, Locust, or `wrk`).

For now, I'll merge this, but if anyone has any idea about how this could be tested, to have checks in CI making sure the problem doesn't happen again, it would be awesome.

Thanks @adriangb for all the investigation and work! :bow: :nerd_face: :cake: