@adriangb 2022-09-04
> I wish there was a way to test this, in pytest. But not sure how to replicate the whole problem and situation in a feasible way (not involving a whole setup for "benchmarks" with ab, Locust, or wrk)

It might be reproducible by talking pure-ASGI to the FastAPI app and launching multiple tasks using a TaskGroup? Might be hard to get the timing right though.

It might be worth trying to set the capacity limit to float("inf") instead of 1? Could the slowdown come from acquiring database connections? There's likely some semaphore there as well.