@tiangolo 2022-09-04
Hehe I was writing this and you commented before me :sweat_smile: 

---

Ah, wait, I didn't read @Kludex comment properly. I didn't realize it could be an inf float. That probably makes sense and we can avoid creating a new object every time.

---

I just di that and updated the branch. :rocket: