@adriangb 2022-09-04
> But I'm not sure how to make it fail instead of hang, if it just hangs it's gonna be more difficult to test

I think anyio.fail_on_after(<some reasonable amount of time>) should do the trick right? But if we were to put this in a test it might be good to figure out how to reproduce it without SQLAlchemy.